refactor: useFilterValue 훅 분리 (mina-gwak/issue-tracker#69)
