+++ FE/src/components/IssueTable/TableHeader/Filter.tsx
@@ -1,6 +1,5 @@
-import { useRef, useState } from 'react';
+import { useRef } from 'react';
 
-import axios from 'axios';
 import { useRecoilValue } from 'recoil';
 
 import * as S from '@components/IssueTable/TableHeader/TableHeader.style';
@@ -8,11 +7,9 @@ import Modal from '@components/Modal';
 import { POSITION } from '@components/Modal/constants';
 import Icon from '@components/common/Icon';
 import { ICON_NAME } from '@components/common/Icon/constants';
-import { API } from '@constants';
-import { stateModify } from '@data/dropdownData';
+import useFilterValue from '@hooks/useFilterValue';
 import { useModal } from '@hooks/useModal';
 import { modalState, ModalStateType } from '@store/dropdown';
-import { DropdownType } from '@type/dropdownType';
 
 interface FilterPropsType {
   type: ModalStateType;
@@ -21,36 +18,13 @@ interface FilterPropsType {
 
 const Filter = ({ type, title }: FilterPropsType) => {
   const modalRef = useRef<HTMLDivElement>(null);
-  const [filterValue, setFilterValue] = useState<DropdownType[]>();
+  const {filterValue, fetchFilterValue} = useFilterValue({ type, title });
   const modalValue = useRecoilValue(modalState);
   const { toggleModal, handleModalClick } = useModal({ modalRef, type });
 
-  const fetchTypeData = async () => {
-    if (filterValue) return;
-    if (type === 'stateModify') {
-      setFilterValue(stateModify);
-      return;
-    }
-    const accessToken = localStorage.getItem('accessToken')!;
-
-    const response = await axios.get(`${API.FILTER(type)}`, {
-      headers: {
-        Authorization: `Bearer ${JSON.parse(accessToken)}`,
-      },
-    });
-
-    setFilterValue([
-      {
-        optionName: `${title}이 없는 이슈`,
-        value: 'none',
-      },
-      ...response.data[`${type}OutlineResponses`],
-    ]);
-  };
-
   return (
     <>
-      <S.FilterButton onMouseEnter={fetchTypeData} onClick={toggleModal}>
+      <S.FilterButton onMouseEnter={fetchFilterValue} onClick={toggleModal}>
         {title}
         <Icon iconName={ICON_NAME.SELECT} />
       </S.FilterButton>
